+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this branch will be documented in this file.
 
 ## 1.3.2 - Unreleased
 
+### Added
+
+- Added a new `Disable Start Menu Recommended Section` option in the `Features`, suggested be @XMontech1337X #31.
+
 ### Fixes
 
 - Fixed an error showing up when loading installed apps on for the first time #30.